+++ src/main/java/com/patryklikus/publicchat/models/MessageBuilder.java
@@ -3,35 +3,35 @@ package com.patryklikus.publicchat.models;
 
 import java.time.LocalDateTime;
 
-public final class PostBuilder {
+public final class MessageBuilder {
     private Long id;
     private User author;
     private String content;
     private LocalDateTime timestamp;
 
-    private PostBuilder() {
+    private MessageBuilder() {
     }
 
-    public static PostBuilder aMessage() {
-        return new PostBuilder();
+    public static MessageBuilder aMessage() {
+        return new MessageBuilder();
     }
 
-    public PostBuilder withId(Long id) {
+    public MessageBuilder withId(Long id) {
         this.id = id;
         return this;
     }
 
-    public PostBuilder withAuthor(User author) {
+    public MessageBuilder withAuthor(User author) {
         this.author = author;
         return this;
     }
 
-    public PostBuilder withContent(String content) {
+    public MessageBuilder withContent(String content) {
         this.content = content;
         return this;
     }
 
-    public PostBuilder withTimestamp(LocalDateTime timestamp) {
+    public MessageBuilder withTimestamp(LocalDateTime timestamp) {
         this.timestamp = timestamp;
         return this;
     }
