+++ src/main/java/com/patryklikus/publicchat/repositories/MessageRepository.java
@@ -1,7 +1,7 @@
 /* Copyright Patryk Likus All Rights Reserved. */
 package com.patryklikus.publicchat.repositories;
 
-import static com.patryklikus.publicchat.models.PostBuilder.aMessage;
+import static com.patryklikus.publicchat.models.MessageBuilder.aMessage;
 import static com.patryklikus.publicchat.models.UserBuilder.anUser;
 
 import com.patryklikus.publicchat.clients.PostgresClient;
